Add solution for balancing parentheses exercise
In this commit, I have implemented the solution 
for the balancing parentheses exercise. The problem
itself can be understood as a matter of delimiter checking,
and I have implemented my solution accordingly:
   
    1. Initialize an empty Stack called `stack`
    2. Using a for loop, iterate over each character in the input
        string and execute the following:
        a. return False if the character is `)` and the stack is empty
        b. push the character to the stack if it `(`
        c. pop the stack if the character is `)`
    3. return True if the stack is empty and False if not

This ensures the parentheses are balanced because if 
anything is left in the stack that means there are too many 
opening parentheses. If the first `if` statement executes then 
that means there are too many closing parentheses.