Implement solution for queue class exercise
In this commit, I implement the solution to the queue
class exercise. In practice, it is similar to a Stack but operates
"first in first out" as opposed to "first in last out". While the exercise 
doesn't test for an empty queue, I made it a point to validate it's not 
empty anyway to get into the habit of accounting for edge cases.