Implement solution for node class exercise
In this commit, I implement the solution to the queue
class exercise. It's fairly straightforward; the constructor 
sets the value of the node to the argument and the reference
to the next node to None, and the `set_next` method sets the
reference to the next node to the argument.