Implement solution for linked list class exercise
In this commit, I implement the solution for the linked
list class exercise. 

The constructor sets the head to None to initialize an empty 
linked list required by the tests.

The iterator method sets the current node to the head of the 
linked list to start, then initializes a while loop that continues
as long as the current node is not None. For each iteration of
the loop, we yield the current node which returns its values and
continues the loop from where it left off, and then we set the current
node to the next value in the linked list. The iterator stops when the 
next value is None which marks the end of the linked list.