Implement solution for add to tail exercise
 In this commit, I have implemented the solution for the 
"add to tail" exercise in a file called `add_to_tail.py`, which 
can be found in the `3_queue`s directory.

The method first checks if the linked list is empty 
and, if so, sets the head to the new node. If the linked 
list is not empty, it initializes an empty variable called 
`last_node` and iterates through the linked list, updating 
`last_node` to the current node. After the iteration, `last_node`
should be pointing to the tail of the linked list, where we can 
finally assign the new node as the next node.