Implement solution for add to head exercise
In this commit I have implemented the solution for the
add to head exercise. The solution itself is straight forward;
if the linked list is empty then the new node becomes the head,
and if not, the new node's next is set to the previous head and
then the new node becomes the head.