Implement solution for linked list queue exercise
In this commit, I implement the solution by introducing a new data
member for the class called `tail`. By saving the end of the list in this
data member we no longer require a loop to reach the end of the list
in the `add_to_tail` method. This changes the method from O(n) to O(1).