Add starter code for matchmake queue exercise
I have implemented starter code for the matchmake queue exercise, in a 
new file named `matchmake_queue.py` in the `3_queue` directory. 

The exercise tasks us to create a function `matchmaking_queue` that simulates 
players joining and leaving the matchmaking queue. The function should take 
a list of player names and their actions (either "join" or "leave") as a tuple.

For each call to matchmake:
1. If the action is "leave", search the queue for the player and remove them if they 
are in the queue.
2. If the action is "join", push the player onto the queue if they are not already in the 
queue.
3. After a push, check if the queue has at least 4 players in it. If so, pop the first 2 players 
from the queue and print the following to the console: "{player1} matched {player2}!"
Where player1 is the first player popped and player2 is the second player popped.