Implement solution for min max exercise
The solution for this exercise uses a while loop to find the left-most
value for get_min` and the right-most value for `get_max`. This solution
works because a proper BST places the smaller value to the left and a
larger value to the right. If we continue to iterate left we end up with
the left-most value, or the smallest value. If we continue right we end
up with the right-most value, or the largest value.