Implement solution for bst delete exercise
The solution works by leveraging several conditional statements to
traverse the BST and validate if each node value is the one
we are looking to delete.

The time complexity of the `delete` method we have created is O(log(n))
because we always halve the number of nodes we are looking at. This is
evident in the code for lines 9-12 and 14-17, which use recursion to
iterate through the left or right branches of the BST depending on
whether the value we are looking to delete is less than or greater than
the current node value. This ensures that we do not have to check every
node in the tree to find the value.

Once we find the value we are looking for, there are three possibilities
we must account for:

1. If there is no right child, we return the left child, which deletes the
current node by bypassing it completely.
2. Similarly, if there is no left child, we return the right child,
effectively deleting the current node.
3. If both the left and right nodes are present, we use a helper method
called `__get_min_larger_val` to find the minimum larger node and
replace the current node with it. (Although a `get_max_smaller_val` method
would also work based on my testing, I followed the exercise instructions
and used the former approach.) Then, we simply replace `self.val` with
`min_larger_val`, and delete `min_larger_val.val` from the right subtree.