Add starter code for hashmap insert exercise
Course: Learn Data Structures
Chapter: 5. Hash Maps
Exercise Link: https://boot.dev/course/7bbb53ed-2106-4f6b-b885-e7645c2ff9d8/d698f2b9-8526-4a11-9e8c-ad45890bd220/fe0bc417-69dc-435e-95c8-79cd8dde04a0

In this exercise we are to complete the insert method which takes a key 
(a string) and a value that can be of any type and stores them in the map 
as a tuple.  We are also to use `key_to_index` to find which index the 
tuple should be stored in.