Add starter code for advanced word filter exercise
Course: Learn Data Structures
Chapter: 6. Tries
Exercise Link: https://boot.dev/course/7bbb53ed-2106-4f6b-b885-e7645c2ff9d8/61e9d21f-63e5-4e9f-b294-4c1972d043ab/dd146e90-6275-448a-89c4-d148e52ab022

In this exercise, we are creating a method `advanced_find_matches` that
will catch variations of bad words in the document by replacing specific
characters.

`advanced_find_matches` takes an entire document and a dictionary of
character variations as input, and should return a set() of all the
words in the trie that exist in the document as continuous substrings,
even if the word had a variation character instead of the original.

For example, if the document contains 'd@rn', and the variations
dictionary contains {'@': 'a'}, and "darn" is in the trie, then "d@rn"
should be returned as a match.